WIP: Removing fixture and simplifying
