+++ docs/examples/microwave/features/cooking_human.feature
@@ -1,6 +1,5 @@
 Feature: Cooking
 
-  @fixture.sismic_application
   Scenario: Start cooking food
     Given I open the door
     And I place an item in the oven
@@ -10,7 +9,6 @@ Feature: Cooking
     When I press start button
     Then heating turns on
 
-  @fixture.sismic_application
   Scenario: Stop cooking food
     Given I reproduce "Start cooking food"
     When 2 seconds elapsed
@@ -19,7 +17,6 @@ Feature: Cooking
     Then variable timer equals 0
     And heating turns off
 
-  @fixture.sismic_application
   Scenario: Cooking stops after preset time
     Given I reproduce "Start cooking food"
     When 5 seconds elapsed
