+++ sismic/bdd/environment.py
@@ -1,4 +1,75 @@
+from sismic.bdd.async_support import async_test, clear_async_context, create_async_context
+from typing import Callable, List
+
+from behave.runner import ModelRunner
 from sismic.helpers import log_trace
+from sismic.interpreter.default import Interpreter
+from sismic.io.yaml import import_from_yaml
+from sismic.model.statechart import Statechart
+from sismic.bdd.steps import *
+
+__all__ = [
+    "setup_behave_context",
+]
+
+
+def setup_behave_context(
+    context,
+    statechart_path=None,
+    statechart: Statechart = None,
+    property_statecharts: List[Statechart] = None,
+    interpreter_klass: Callable[[Statechart], Interpreter] = Interpreter,
+    async_start_fn=None,
+    is_async=False,
+    reset_fn=None,
+):
+    property_statecharts = property_statecharts if property_statecharts else []
+    sc = statechart
+    if statechart_path:
+        sc = import_from_yaml(filepath=statechart_path)
+    if not sc:
+        raise Exception("No statechart found")
+    context.config.update_userdata(
+        {
+            "statechart": sc,
+            "interpreter_klass": interpreter_klass,
+            "is_async": is_async,
+            "property_statecharts": property_statecharts,
+            "async_start_fn": async_start_fn,
+            "reset_fn": reset_fn,
+        }
+    )
+
+    behave_run_hook = ModelRunner.run_hook
+
+    def run_hook(self, name, context, *args):
+        if name == "before_scenario":
+            sismic_before_scenario(context, *args)
+            if async_test(context):
+                if getattr(context, "fixture-setup"):
+                    setattr(context, "fixture-setup", False)
+                else:
+                    create_async_context(context)
+                    sismic_before_scenario(context, *args)
+        elif name == "before_step":
+            sismic_before_step(context, *args)
+
+        behave_run_hook(self, name, context, *args)
+
+        if name == "after_scenario":
+            if async_test(context):
+                clear_async_context(context)
+        elif name == "after_step":
+            sismic_after_step(context, *args)
+        elif name == "after_all":
+            sismic_after_all(context, *args)
+
+    ModelRunner.run_hook = run_hook
+
+
+def sismic_after_all(context):
+    if async_test(context):
+        clear_async_context(context)
 
 
 def setup_sismic_from_context(context):
@@ -73,3 +144,14 @@ def sismic_after_step(context, step):
         print("--------------------------------------------------------------")
 
         pdb.post_mortem(step.exc_traceback)
+
+
+async def default_async_task(context):
+    async_start_fn = context.config.userdata.get("async_start_fn")
+    try:
+        if async_start_fn:
+            await (async_start_fn)(context)
+        setup_sismic_from_context(context)
+        await asyncio.sleep(0)
+    except Exception as ex:
+        print(ex)
